Fix bugs patches 0004 and 0005
Patch 0004 was completely broken on 32-bit systems, while patch 0005
didn’t actually do what it was supposed to.  There is no security impact
because the “out-of-bounds” read would actually be from the header
prelude (which is from the same allocation as the rest of the header)
and the package would be rejected immediately afterwards.